fix: Making classes accessible to windows via uno.ui.toolkit
