+++ src/Uno.Foundation/Uno.Core.Extensions/Uno.Core.Extensions.Collections/EnumerableExtensions.cs
@@ -18,8 +18,6 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-using Uno.Collections;
-using System.Collections.ObjectModel;
 
 namespace Uno.Extensions
 {
