+++ src/Uno.UI.RemoteControl/HotReload/ClientHotReloadProcessor.MetadataUpdate.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable enable
+
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
@@ -10,32 +12,29 @@ using System.Threading.Tasks;
 using Uno.Extensions;
 using Uno.Foundation.Logging;
 using Uno.UI.Helpers;
-using Uno.UI.RemoteControl.HotReload.Messages;
 using Uno.UI.RemoteControl.HotReload.MetadataUpdater;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 
 namespace Uno.UI.RemoteControl.HotReload;
 
-partial class ClientHotReloadProcessor : IRemoteControlProcessor
+partial class ClientHotReloadProcessor
 {
 	private static int _isReloading;
 
-	private bool _linkerEnabled;
-	private HotReloadAgent _agent;
-	private ElementUpdateAgent? _elementAgent;
-	private static ClientHotReloadProcessor? _instance;
-	private readonly TaskCompletionSource<bool> _hotReloadWorkloadSpaceLoaded = new();
+	private static ElementUpdateAgent? _elementAgent;
 
-	private ElementUpdateAgent ElementAgent
+	private static Logger _log = typeof(ClientHotReloadProcessor).Log();
+
+	private static ElementUpdateAgent ElementAgent
 	{
 		get
 		{
 			_elementAgent ??= new ElementUpdateAgent(s =>
 				{
-					if (this.Log().IsEnabled(LogLevel.Trace))
+					if (_log.IsEnabled(LogLevel.Trace))
 					{
-						this.Log().Trace(s);
+						_log.Trace(s);
 					}
 				});
 
@@ -44,15 +43,13 @@ partial class ClientHotReloadProcessor : IRemoteControlProcessor
 	}
 
 	private void WorkspaceLoadResult(HotReloadWorkspaceLoadResult hotReloadWorkspaceLoadResult)
-		=> _hotReloadWorkloadSpaceLoaded.SetResult(hotReloadWorkspaceLoadResult.WorkspaceInitialized);
+		=> _hotreloadWorkloadSpaceLoaded.SetResult(hotReloadWorkspaceLoadResult.WorkspaceInitialized);
 
 	/// <summary>
-	/// Waits for the server's hot reload workspace to be loaded
+	/// Determines if the server's hot reload workspace has been loaded
 	/// </summary>
-	/// <param name="ct"></param>
-	/// <returns></returns>
-	public Task<bool> WaitForWorkspaceLoaded(CancellationToken ct)
-		=> _hotReloadWorkloadSpaceLoaded.Task;
+	internal Task HotReloadWorkspaceLoaded
+		=> _hotreloadWorkloadSpaceLoaded.Task;
 
 	[MemberNotNull(nameof(_agent))]
 	partial void InitializeMetadataUpdater()
@@ -76,140 +73,6 @@ partial class ClientHotReloadProcessor : IRemoteControlProcessor
 				this.Log().Trace(s);
 			}
 		});
-	}
-
-	private bool MetadataUpdatesEnabled
-	{
-		get
-		{
-			var unoRuntimeIdentifier = GetMSBuildProperty("UnoRuntimeIdentifier");
-			var targetFramework = GetMSBuildProperty("TargetFramework");
-			var buildingInsideVisualStudio = GetMSBuildProperty("BuildingInsideVisualStudio");
-
-			return
-				buildingInsideVisualStudio.Equals("true", StringComparison.OrdinalIgnoreCase)
-				&& (
-					// As of VS 17.8, when the debugger is not attached, mobile targets can use
-					// DevServer's hotreload workspace, as visual studio does not enable it on its own.
-					(!Debugger.IsAttached
-						&& (targetFramework.Contains("-android") || targetFramework.Contains("-ios"))));
-		}
-	}
-
-	private string[] GetMetadataUpdateCapabilities()
-	{
-		if (Type.GetType(HotReloadAgent.MetadataUpdaterType) is { } type)
-		{
-			if (type.GetMethod("GetCapabilities", BindingFlags.Static | BindingFlags.NonPublic) is { } getCapabilities)
-			{
-				if (getCapabilities.Invoke(null, Array.Empty<string>()) is string caps)
-				{
-					if (this.Log().IsEnabled(LogLevel.Trace))
-					{
-						this.Log().Trace($"Metadata Updates runtime capabilities: {caps}");
-					}
-
-					return caps.Split(' ');
-				}
-				else
-				{
-					if (this.Log().IsEnabled(LogLevel.Warning))
-					{
-						this.Log().Trace($"Runtime does not support Hot Reload (Invalid returned type for {HotReloadAgent.MetadataUpdaterType}.GetCapabilities())");
-					}
-				}
-			}
-			else
-			{
-				if (this.Log().IsEnabled(LogLevel.Warning))
-				{
-					this.Log().Trace($"Runtime does not support Hot Reload (Unable to find method {HotReloadAgent.MetadataUpdaterType}.GetCapabilities())");
-				}
-			}
-		}
-		else
-		{
-			if (this.Log().IsEnabled(LogLevel.Warning))
-			{
-				this.Log().Trace($"Runtime does not support Hot Reload (Unable to find type {HotReloadAgent.MetadataUpdaterType})");
-			}
-		}
-		return Array.Empty<string>();
-	}
-
-	private void AssemblyReload(AssemblyDeltaReload assemblyDeltaReload)
-	{
-		try
-		{
-			if (Debugger.IsAttached)
-			{
-				if (this.Log().IsEnabled(LogLevel.Error))
-				{
-					this.Log().Error($"Hot Reload is not supported when the debugger is attached.");
-				}
-
-				return;
-			}
-
-			if (assemblyDeltaReload.IsValid())
-			{
-				if (this.Log().IsEnabled(LogLevel.Trace))
-				{
-					this.Log().Trace($"Applying IL Delta after {assemblyDeltaReload.FilePath}, Guid:{assemblyDeltaReload.ModuleId}");
-				}
-
-				var changedTypesStreams = new MemoryStream(Convert.FromBase64String(assemblyDeltaReload.UpdatedTypes));
-				var changedTypesReader = new BinaryReader(changedTypesStreams);
-
-				var delta = new UpdateDelta()
-				{
-					MetadataDelta = Convert.FromBase64String(assemblyDeltaReload.MetadataDelta),
-					ILDelta = Convert.FromBase64String(assemblyDeltaReload.ILDelta),
-					PdbBytes = Convert.FromBase64String(assemblyDeltaReload.PdbDelta),
-					ModuleId = Guid.Parse(assemblyDeltaReload.ModuleId),
-					UpdatedTypes = ReadIntArray(changedTypesReader)
-				};
-
-				_agent.ApplyDeltas(new[] { delta });
-
-				if (this.Log().IsEnabled(LogLevel.Trace))
-				{
-					this.Log().Trace($"Done applying IL Delta for {assemblyDeltaReload.FilePath}, Guid:{assemblyDeltaReload.ModuleId}");
-				}
-			}
-			else
-			{
-				if (this.Log().IsEnabled(LogLevel.Trace))
-				{
-					this.Log().Trace($"Failed to apply IL Delta for {assemblyDeltaReload.FilePath} ({assemblyDeltaReload})");
-				}
-			}
-		}
-		catch (Exception e)
-		{
-			if (this.Log().IsEnabled(LogLevel.Error))
-			{
-				this.Log().Error($"An exception occurred when applying IL Delta for {assemblyDeltaReload.FilePath} ({assemblyDeltaReload.ModuleId})", e);
-			}
-		}
-	}
-
-	static int[] ReadIntArray(BinaryReader binaryReader)
-	{
-		var numValues = binaryReader.ReadInt32();
-		if (numValues == 0)
-		{
-			return Array.Empty<int>();
-		}
-
-		var values = new int[numValues];
-
-		for (var i = 0; i < numValues; i++)
-		{
-			values[i] = binaryReader.ReadInt32();
-		}
-
-		return values;
 	}
 
 	private static async Task<bool> ShouldReload()
@@ -240,7 +103,7 @@ partial class ClientHotReloadProcessor : IRemoteControlProcessor
 		{
 			UpdateGlobalResources(updatedTypes);
 
-			var handlerActions = _instance?.ElementAgent?.ElementHandlerActions;
+			var handlerActions = ElementAgent?.ElementHandlerActions;
 
 			// Action: BeforeVisualTreeUpdate
 			// This is called before the visual tree is updated
@@ -486,12 +349,12 @@ partial class ClientHotReloadProcessor : IRemoteControlProcessor
 			switch (instance)
 			{
 #if __IOS__
-				case UserControl userControl:
-					if (newInstance is UIKit.UIView newUIViewContent)
-					{
-						SwapViews(userControl, newUIViewContent);
-					}
-					break;
+			case UserControl userControl:
+				if (newInstance is UIKit.UIView newUIViewContent)
+				{
+					SwapViews(userControl, newUIViewContent);
+				}
+				break;
 #endif
 				case ContentControl content:
 					if (newInstance is ContentControl newContent)
