+++ src/Uno.UI/Helpers/ElementMetadataUpdateHandlerAttribute.cs
@@ -1,24 +1,22 @@
-﻿namespace System.Reflection.Metadata
+﻿namespace System.Reflection.Metadata;
+
+/// <summary>
+/// Defines a handler for an type of element that is found in the visual tree
+/// which will be invokved if there are metadata updates
+/// </summary>
+[AttributeUsage(AttributeTargets.Assembly, AllowMultiple = true)]
+public sealed class ElementMetadataUpdateHandlerAttribute : Attribute
 {
-	/// <summary>
-	/// Defines a handler for an type of element that is found in the visual tree
-	/// which will be invokved if there are metadata updates
-	/// </summary>
-	[AttributeUsage(AttributeTargets.Assembly, AllowMultiple = true)]
-	public sealed class ElementMetadataUpdateHandlerAttribute : Attribute
+	public Type ElementType { get; }
+	public Type HandlerType { get; }
+
+	public ElementMetadataUpdateHandlerAttribute(Type elementType, Type handlerType)
 	{
-		public Type ElementType { get; }
-		public Type HandlerType { get; }
-
-		public ElementMetadataUpdateHandlerAttribute(Type elementType, Type handlerType)
-		{
-			ElementType = elementType;
-			HandlerType = handlerType;
-		}
-
-		public ElementMetadataUpdateHandlerAttribute(Type handlerType) : this(typeof(object), handlerType)
-		{
-		}
+		ElementType = elementType;
+		HandlerType = handlerType;
 	}
 
+	public ElementMetadataUpdateHandlerAttribute(Type handlerType) : this(typeof(object), handlerType)
+	{
+	}
 }
